[meta] arranging some nodes
