+++ chain/build-dictionary.go
@@ -5,12 +5,6 @@ import (
 	"sync"
 )
 
-// we're hard-coding order of 2 markov chains here
-type bigram struct {
-	First, Second string
-}
-type chain map[bigram][]string
-
 func chainFromTweet(tweet string, pointer *chain, pool *sync.WaitGroup) {
 	tokens := strings.Split(tweet, " ")
 	tweetChain := make(chain)
@@ -56,21 +50,4 @@ func buildDictionary(tweets []string) chain {
 	}
 
 	return markovChain
-}
-
-// Use the pair of words at the beginning of each tweet as a set of
-// potentional starters
-func findStarters(tweets []string) []bigram {
-	starters := []bigram{}
-
-	for _, tweet := range tweets {
-		words := strings.Split(tweet, " ")
-
-		if len(words) > 2 {
-			pair := bigram{words[0], words[1]}
-			starters = append(starters, pair)
-		}
-	}
-
-	return starters
 }
